refactor: create task manager file and move logic there
