+++ examples/incredible-dot-product/contracts/script/deployments/incredible-dot-product/31337.json
@@ -1,1 +1,1 @@
-{"lastUpdate":{"timestamp":"1746650926","block_number":"22"},"addresses":{"proxyAdmin":"0x1291be112d480055dafd8a610b7d1e203891c274","incredibleDotProductServiceManager":"0xcd8a1c3ba11cf5ecfa6267617243239504a98d90","incredibleDotProductServiceManagerImpl":"0x202cce504e04bed6fc0521238ddf04bc9e8e15ab","incredibleDotProductTaskManager":"0x7bc06c482dead17c0e297afbc32f6e63d3846650","registryCoordinator":"0xfd471836031dc5108809d173a067e8486b9047a3","blsapkRegistry":"0x1429859428c0abc9c2c47c8ee9fbaf82cfa0f20f","indexRegistry":"0x162a433068f51e18b7d13932f27e66a3f99e6890","stakeRegistry":"0x2bdcc0de6be1f7d2ee689a0342d76f52e8efaba3","operatorStateRetriever":"0x5f3f1dbd7b74c6b46e8c44f98792a1daf8d69154","strategy":"0x2b961e3959b79326a8e7f64ef0d2d825707669b5","pauserRegistry":"0x0000000000000000000000000000000000000000","token":"0x9d4454b023096f34b160d6b654540c56a1f81688","instantSlasher":"0xdbc43ba45381e02825b14322cddd15ec4b3164e6"}}
+{"lastUpdate":{"timestamp":"1746651500","block_number":"38"},"addresses":{"proxyAdmin":"0x1291be112d480055dafd8a610b7d1e203891c274","incredibleDotProductServiceManager":"0xcd8a1c3ba11cf5ecfa6267617243239504a98d90","incredibleDotProductServiceManagerImpl":"0x202cce504e04bed6fc0521238ddf04bc9e8e15ab","incredibleDotProductTaskManager":"0x7bc06c482dead17c0e297afbc32f6e63d3846650","registryCoordinator":"0xfd471836031dc5108809d173a067e8486b9047a3","blsapkRegistry":"0x1429859428c0abc9c2c47c8ee9fbaf82cfa0f20f","indexRegistry":"0x162a433068f51e18b7d13932f27e66a3f99e6890","stakeRegistry":"0x2bdcc0de6be1f7d2ee689a0342d76f52e8efaba3","operatorStateRetriever":"0x5f3f1dbd7b74c6b46e8c44f98792a1daf8d69154","strategy":"0x2b961e3959b79326a8e7f64ef0d2d825707669b5","pauserRegistry":"0x0000000000000000000000000000000000000000","token":"0x9d4454b023096f34b160d6b654540c56a1f81688","instantSlasher":"0xdbc43ba45381e02825b14322cddd15ec4b3164e6"}}
